chore: Split sketch generation methods out of sample_sketches and into testing_utils.sketch_gen
